refactor(menuPriceValidator)
