+++ src/main/java/kitchenpos/menus/tobe/domain/MenuPriceValidator.kt
@@ -2,4 +2,6 @@ package kitchenpos.menus.tobe.domain
 
 interface MenuPriceValidator {
     fun validate(menu: Menu)
+
+    fun isValid(menu: Menu): Boolean
 }
