+++ src/test/java/kitchenpos/menus/application/tobe/FakeMenuPriceValidator.kt
@@ -5,12 +5,16 @@ import kitchenpos.menus.tobe.domain.Menu
 import kitchenpos.menus.tobe.domain.MenuPriceValidator
 import java.math.BigDecimal
 
-val FAKE_MENU_PRODUCTS_TOTAL_PRICES = BigDecimal.valueOf(500).price()
+val FAKE_MENU_PRODUCTS_TOTAL_PRICES = BigDecimal.valueOf(50000).price()
 
 object FakeMenuPriceValidator : MenuPriceValidator {
     override fun validate(menu: Menu) {
-        if (menu.price <= FAKE_MENU_PRODUCTS_TOTAL_PRICES) {
+        if (menu.price > FAKE_MENU_PRODUCTS_TOTAL_PRICES) {
             throw IllegalArgumentException("메뉴상품 가격의 총합보다 메뉴의 가격이 클 수 없습니다")
         }
+    }
+
+    override fun isValid(menu: Menu): Boolean {
+        return menu.price <= FAKE_MENU_PRODUCTS_TOTAL_PRICES
     }
 }
